[From Anna Sokol] Fixed root for cameras using RELATIVE_RF.
The difference between cv->getModelViewMatrix() and
cv->getCurrentCamera()->getViewMatrix() only exists when the main
camera is using RELATIVE_RF (reference frame) because in that instance
the two matrices above get multiplied to make a new modelViewMatrix
for the cullVisitor. In an ABSOLUTE_RF
the cv->getModelViewMatrix() is a direct
cv->getCurrentCamera()->getViewMatrix().

This all matters because when there is an object with sound
in a scene, the scene view doesn't start from 0,0,0, and the main
camera uses RELATIVE_RF, the sound doesn't come from the location it
should.